requirements.txt
attrs>=22.1.0  # Python class wrapper
matplotlib>=3.6.0  # Displaying the UKF results
PyHamcrest>=2.0.4  # Testing frameworks
pylint>=2.15.7 # Code analysis
pytest>=7.2.0  # Testing frameworks
sphinx>=5.3.0  # Python documentation generator
